refactor: Move recipe file to services
